@@ -4,14 +4,15 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.13.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  isolate_bloc:
-    path: ../
+  isolate_bloc: ^2.0.0
+  # isolate_bloc:
+    # path: ../
 
 dev_dependencies:
   flutter_test:
