@@ -1,11 +1,12 @@
 name: isolate_bloc_test_example
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
-  flutter: ">=1.17.0 <2.0.0"
+  sdk: ">=2.12.0 <3.0.0"
+  flutter: ">=2.0.0 <3.0.0"
 
 dependencies:
-  isolate_bloc: ^0.4.0
+  isolate_bloc: 
+    path: ../../isolate_bloc/
   isolate_bloc_test:
     path: ../
   flutter:
