@@ -1,5 +1,4 @@
 import 'package:equatable/equatable.dart';
-import 'package:flutter/foundation.dart';
 import 'package:isolate_bloc/isolate_bloc.dart';
 import 'package:weather_app/models/weather.dart';
 import 'package:weather_app/repositories/repositories.dart';
@@ -7,36 +6,35 @@ import 'package:weather_app/repositories/repositories.dart';
 class WeatherBloc extends IsolateBloc<WeatherEvent, WeatherState> {
   final WeatherRepository weatherRepository;
 
-  WeatherBloc({@required this.weatherRepository})
-      : assert(weatherRepository != null),
-        super(WeatherInitial());
+  WeatherBloc({required this.weatherRepository}) : super(WeatherInitial());
+
+  Stream<WeatherState> _weatherRequested(WeatherRequested event) async* {
+    yield WeatherLoadInProgress();
+    try {
+      final Weather weather = await weatherRepository.getWeather(event.city);
+      yield WeatherLoadSuccess(weather: weather);
+    } catch (e) {
+      print(e);
+      yield WeatherLoadFailure();
+    }
+  }
+
+  Stream<WeatherState> _weatherRefreshRequested(
+      WeatherRefreshRequested event) async* {
+    try {
+      final Weather weather = await weatherRepository.getWeather(event.city);
+      yield WeatherLoadSuccess(weather: weather);
+    } catch (e) {
+      print(e);
+    }
+  }
 
   @override
-  void onEventReceived(WeatherEvent event) {
+  Stream<WeatherState> mapEventToState(WeatherEvent event) async* {
     if (event is WeatherRequested) {
-      _weatherRequested(event);
+      yield* _weatherRequested(event);
     } else if (event is WeatherRefreshRequested) {
-      _weatherRefreshRequested(event);
-    }
-  }
-
-  Future<void> _weatherRequested(WeatherRequested event) async {
-    emit(WeatherLoadInProgress());
-    try {
-      final Weather weather = await weatherRepository.getWeather(event.city);
-      emit(WeatherLoadSuccess(weather: weather));
-    } catch (e) {
-      print(e);
-      emit(WeatherLoadFailure());
-    }
-  }
-
-  Future<void> _weatherRefreshRequested(WeatherRefreshRequested event) async {
-    try {
-      final Weather weather = await weatherRepository.getWeather(event.city);
-      emit(WeatherLoadSuccess(weather: weather));
-    } catch (e) {
-      print(e);
+      yield* _weatherRefreshRequested(event);
     }
   }
 }
@@ -48,7 +46,7 @@ abstract class WeatherEvent extends Equatable {
 class WeatherRequested extends WeatherEvent {
   final String city;
 
-  const WeatherRequested({@required this.city}) : assert(city != null);
+  const WeatherRequested({required this.city});
 
   @override
   List<Object> get props => [city];
@@ -57,7 +55,7 @@ class WeatherRequested extends WeatherEvent {
 class WeatherRefreshRequested extends WeatherEvent {
   final String city;
 
-  const WeatherRefreshRequested({@required this.city}) : assert(city != null);
+  const WeatherRefreshRequested({required this.city});
 
   @override
   List<Object> get props => [city];
@@ -77,7 +75,7 @@ class WeatherLoadInProgress extends WeatherState {}
 class WeatherLoadSuccess extends WeatherState {
   final Weather weather;
 
-  const WeatherLoadSuccess({@required this.weather}) : assert(weather != null);
+  const WeatherLoadSuccess({required this.weather});
 
   @override
   List<Object> get props => [weather];
