@@ -1,17 +1,14 @@
 import 'package:flutter/material.dart';
 
-import 'package:meta/meta.dart';
-
 class GradientContainer extends StatelessWidget {
   final Widget child;
   final MaterialColor color;
 
   const GradientContainer({
-    Key key,
-    @required this.color,
-    @required this.child,
-  })  : assert(color != null, child != null),
-        super(key: key);
+    Key? key,
+    required this.color,
+    required this.child,
+  }) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -22,9 +19,9 @@ class GradientContainer extends StatelessWidget {
           end: Alignment.bottomCenter,
           stops: [0.6, 0.8, 1.0],
           colors: [
-            color[700],
-            color[500],
-            color[300],
+            color[700]!,
+            color[500]!,
+            color[300]!,
           ],
         ),
       ),
