@@ -1,142 +1,146 @@
 import 'dart:async';
+import 'dart:collection';
 
 import 'package:flutter/foundation.dart';
-
-import '../isolate/service_events.dart';
-import 'isolate_bloc.dart';
-import 'transition.dart';
+import 'package:isolate_bloc/isolate_bloc.dart';
+import 'package:isolate_bloc/src/common/isolate/isolate_bloc_events/isolate_bloc_events.dart';
+import 'package:uuid/uuid.dart';
 
 /// Signature for event receiver function which takes an [IsolateBlocTransitionEvent]
 /// and send this event to the [IsolateBloc]
-typedef EventReceiver = void Function(IsolateBlocTransitionEvent<Object> event);
+typedef EventReceiver = void Function(Object? event);
 
 /// Signature for function which takes [IsolateBloc]'s uuid and close it
 typedef IsolateBlocKiller = void Function(String uuid);
 
-/// [IsolateBlocWrapper] work like a client for [IsolateBloc]. It receives [IsolateBloc]'s
-/// states and send events added by `wrapperInstance.add(YourEvent())`. So you can
+/// Takes a `Stream` of `Events` as input
+/// and transforms them into a `Stream` of `States` as output using [IsolateBlocBase].
+///
+/// It works like a client for [IsolateBlocBase]. It receives [IsolateBlocBase]'s
+/// states and sends events added by `wrapperInstance.add(YourEvent())`. So you can
 /// listen for origin bloc's state with `wrapperInstance.listen((state) { })` and add
 /// events as shown above.
-/// createBloc function create [IsolateBloc] in [Isolate] and return this object.
-class IsolateBlocWrapper<State> extends Stream<State> implements Sink<Object> {
-  /// Receives initialState, function which receive events and send them to the
-  /// origin [IsolateBloc] and function which called in [close] and close origin bloc.
-  IsolateBlocWrapper(
+///
+/// It may be created:
+///   * by [createBloc] function which creates [IsolateBlocBase] in [Isolate]
+///     and returns the instance of this class.
+///   * by [getBloc] function which creates the instance of this class
+///     and connects it to the [IsolateBlocBase]
+///
+/// Don't create this manually!
+class IsolateBlocWrapper<State> implements Sink<Object?> {
+  /// Takes initialState ([state]), function which receives events
+  /// and sends them to the [IsolateBlocBase]
+  /// and function which called on [close] and closes [IsolateBlocBase]
+  /// which is connected to this wrapper.
+  @protected
+  IsolateBlocWrapper({
+    State? state,
+    required EventReceiver eventReceiver,
+    required IsolateBlocKiller onBlocClose,
+  })  : _eventReceiver = eventReceiver,
+        _onBlocClose = onBlocClose,
+        _state = state,
+        isolateBlocId = isolateBlocIdGenerator() {
+    _bindEventsListener();
+  }
+
+  /// Creates wrapper for [getBloc] functionality
+  @protected
+  IsolateBlocWrapper.isolate(
+    this._eventReceiver,
+    this._onBlocClose, [
     this._state,
-    this._eventReceiver,
-    this._onBlocClose,
-  ) : _initStateProvided = true {
+  ]) {
     _bindEventsListener();
   }
-
-  /// Create object as default constructor do but without initialState.
-  IsolateBlocWrapper.noInitState(
-    this._eventReceiver,
-    this._onBlocClose,
-  )   : _state = null,
-        _initStateProvided = false {
-    _bindEventsListener();
-  }
-
-  final _eventController = StreamController<Object>.broadcast();
-  final _stateController = StreamController<State>.broadcast();
 
   /// Id of IsolateBloc. It's needed to find bloc in isolate.
-  String _originBlocUuid;
+  ///
+  /// This id may be changed
+  @protected
+  String? isolateBlocId;
 
-  State _state;
-  bool _initStateProvided;
-  final _unsentEvents = <Object>[];
+  final _eventController = StreamController<Object?>.broadcast();
+  final _stateController = StreamController<State>.broadcast();
+
+  State? _state;
+
+  /// Used to sync unsent events
+  var _blocCreated = false;
+  final _unsentEvents = Queue<Object?>();
   final IsolateBlocKiller _onBlocClose;
-  StreamSubscription<Transition<Object, State>> _stateTransitionSubscription;
-  StreamSubscription<Object> _eventReceiverSubscription;
+  late StreamSubscription<Object?> _eventReceiverSubscription;
 
-  /// Returns stream with `event`
-  Stream<Object> get eventStream => _eventController.stream;
-
-  /// Callback which receive events and send them to the IsolateBloc
+  /// Callback which receives events and sends them to the IsolateBloc
   final EventReceiver _eventReceiver;
 
   /// Returns the current [state] of the [bloc].
-  State get state => _state;
+  ///
+  /// It may be null only in wrapper provided by `getBlocWrapperFunction`
+  /// Can't be null in UI isolate
+  State? get state => _state;
 
-  /// Returns whether the `Stream<State>` is a broadcast stream.
-  @override
-  bool get isBroadcast => _stateController.stream.isBroadcast;
+  /// Returns the stream of states
+  Stream<State> get stream => _stateController.stream;
 
-  /// Adds a subscription to the `Stream<State>`.
-  /// Returns a [StreamSubscription] which handles events from
-  /// the `Stream<State>` using the provided [onData], [onError] and [onDone]
-  /// handlers.
-  @override
-  StreamSubscription<State> listen(
-    void Function(State) onData, {
-    Function onError,
-    void Function() onDone,
-    bool cancelOnError,
-  }) {
-    return _prepareStateStream.listen(
-      onData,
-      onError: onError,
-      onDone: onDone,
-      cancelOnError: cancelOnError,
-    );
-  }
-
-  Stream<State> get _prepareStateStream async* {
-    if (_initStateProvided) {
-      yield state;
-    }
-    yield* _stateController.stream;
-  }
+  /// Returns stream of `event`
+  Stream<Object?> get _eventStream => _eventController.stream;
 
   /// As a result, call original [IsolateBloc]'s add function.
   @override
-  void add(Object event) {
+  void add(Object? event) {
     _eventController.add(event);
   }
 
-  /// Closes the `event` stream and request to close [IsolateBloc]
+  /// Closes the `event` stream and requests to close connected [IsolateBlocBase]
   @override
   @mustCallSuper
   Future<void> close() async {
-    _onBlocClose(_originBlocUuid);
+    var id = isolateBlocId;
+    if (id != null) {
+      _onBlocClose(id);
+    }
     await _eventController.close();
     await _stateController.close();
-    await _stateTransitionSubscription?.cancel();
-    await _eventReceiverSubscription?.cancel();
+    await _eventReceiverSubscription.cancel();
   }
 
-  /// Connect this wrapper to the origin [IsolateBloc] and start listening for state.
-  void connectToBloc(String uuid) {
-    assert(uuid != null);
-    _originBlocUuid = uuid;
+  /// Connects this wrapper to the [IsolateBlocBase] and sends all unsent events.
+  // todo(maksim): maybe move unsent events synchronization to the [IsolateManager]
+  @protected
+  void onBlocCreated() {
+    _blocCreated = true;
     while (_unsentEvents.isNotEmpty) {
-      _eventReceiver(IsolateBlocTransitionEvent<Object>(
-        _originBlocUuid,
-        _unsentEvents.removeAt(0),
-      ));
+      _eventReceiver(_unsentEvents.removeFirst());
     }
   }
 
-  /// Receive [IsolateBloc]'s state and add to the state Stream.
+  /// Receives [IsolateBlocBase]'s states and adds them to the state Stream.
+  @protected
   void stateReceiver(State nextState) {
-    _initStateProvided = true;
-    if (nextState != state) {
+    if (nextState != _state) {
       _stateController.add(nextState);
+      _state = nextState;
     }
-    _state = nextState;
   }
 
-  /// Start listening for new `events`
+  /// Starts listening for new `events`
   void _bindEventsListener() {
-    _eventReceiverSubscription = eventStream.listen((event) {
-      if (_originBlocUuid != null) {
-        _eventReceiver(
-            IsolateBlocTransitionEvent<Object>(_originBlocUuid, event));
+    _eventReceiverSubscription = _eventStream.listen((event) {
+      if (_blocCreated) {
+        _eventReceiver(event);
       } else {
         _unsentEvents.add(event);
       }
     });
   }
 }
+
+/// Signature for [IsolateBlocWrapper] id generator
+typedef IdGenerator = String Function();
+
+/// This function is used to generate id for [IsolateBlocWrapper]
+///
+/// By default uses `uuid v4` generator
+IdGenerator isolateBlocIdGenerator = const Uuid().v4;
