@@ -1,33 +1,30 @@
-import 'dart:async';
-
 import 'package:flutter/widgets.dart';
-
-import '../common/bloc/isolate_bloc.dart';
-import '../common/bloc/isolate_bloc_wrapper.dart';
-import './isolate_bloc_provider.dart';
+import 'package:isolate_bloc/src/common/bloc/isolate_bloc_base.dart';
+import 'package:isolate_bloc/src/common/bloc/isolate_bloc_wrapper.dart';
+import 'package:isolate_bloc/src/flutter/isolate_bloc_listener.dart';
+import 'package:isolate_bloc/src/flutter/isolate_bloc_provider.dart';
 
 /// Signature for the `builder` function which takes the `BuildContext` and
 /// [state] and is responsible for returning a widget which is to be rendered.
 /// This is analogous to the `builder` function in [StreamBuilder].
-typedef IsolateBlocWidgetBuilder<S> = Widget Function(
-    BuildContext context, S state);
+typedef BlocWidgetBuilder<S> = Widget Function(BuildContext context, S state);
 
 /// Signature for the `buildWhen` function which takes the previous `state` and
 /// the current `state` and is responsible for returning a [bool] which
 /// determines whether to rebuild [IsolateBlocBuilder] with the current `state`.
-typedef IsolateBlocBuilderCondition<S> = bool Function(S previous, S current);
+typedef BlocBuilderCondition<S> = bool Function(S previous, S current);
 
 /// {@template bloc_builder}
 /// [IsolateBlocBuilder] handles building a widget in response to new `states`.
 /// [IsolateBlocBuilder] is analogous to [StreamBuilder] but has simplified API to
-/// reduce the amount of boilerplate code needed as well as [isolateBloc]-specific
+/// reduce the amount of boilerplate code needed as well as [bloc]-specific
 /// performance improvements.
 
-/// Please refer to `BlocListener` if you want to "do" anything in response to
+/// Please refer to [IsolateBlocListener] if you want to "do" anything in response to
 /// `state` changes such as navigation, showing a dialog, etc...
 ///
-/// If the [isolateBloc] parameter is omitted, [IsolateBlocBuilder] will automatically
-/// perform a lookup using [IsolateBlocProvider] and the current `BuildContext`.
+/// If the [bloc] parameter is omitted, [IsolateBlocBuilder] will automatically
+/// perform a lookup using [IsolateBlocProvider] and the current [BuildContext].
 ///
 /// ```dart
 /// IsolateBlocBuilder<BlocA, BlocAState>(
@@ -37,26 +34,29 @@ typedef IsolateBlocBuilderCondition<S> = bool Function(S previous, S current);
 /// )
 /// ```
 ///
-/// Only specify the [isolateBloc] if you wish to provide a [isolateBloc] that is otherwise
-/// not accessible via [IsolateBlocProvider] and the current `BuildContext`.
+/// Only specify the [bloc] if you wish to provide a [bloc] that is otherwise
+/// not accessible via [IsolateBlocProvider] and the current [BuildContext].
 ///
 /// ```dart
 /// IsolateBlocBuilder<BlocA, BlocAState>(
-///   cubit: blocA,
+///   bloc: blocA,
 ///   builder: (context, state) {
 ///   // return widget here based on BlocA's state
 ///   }
 /// )
 /// ```
 /// {@endtemplate}
+///
 /// {@template bloc_builder_build_when}
 /// An optional [buildWhen] can be implemented for more granular control over
 /// how often [IsolateBlocBuilder] rebuilds.
-/// [buildWhen] will be invoked on each [isolateBloc] `state` change.
+/// [buildWhen] should only be used for performance optimizations as it
+/// provides no security about the state passed to the [builder] function.
+/// [buildWhen] will be invoked on each [bloc] `state` change.
 /// [buildWhen] takes the previous `state` and current `state` and must
 /// return a [bool] which determines whether or not the [builder] function will
 /// be invoked.
-/// The previous `state` will be initialized to the `state` of the [isolateBloc] when
+/// The previous `state` will be initialized to the `state` of the [bloc] when
 /// the [IsolateBlocBuilder] is initialized.
 /// [buildWhen] is optional and if omitted, it will default to `true`.
 ///
@@ -69,25 +69,25 @@ typedef IsolateBlocBuilderCondition<S> = bool Function(S previous, S current);
 ///   builder: (context, state) {
 ///     // return widget here based on BlocA's state
 ///   }
-///)
+/// )
 /// ```
 /// {@endtemplate}
-class IsolateBlocBuilder<C extends IsolateBloc<Object, S>, S>
-    extends IsolateBlocBuilderBase<C, S> {
+class IsolateBlocBuilder<B extends IsolateBlocBase<Object?, S>, S>
+    extends IsolateBlocBuilderBase<B, S> {
   /// {@macro bloc_builder}
+  /// {@macro bloc_builder_build_when}
   const IsolateBlocBuilder({
-    Key key,
-    @required this.builder,
-    IsolateBlocWrapper<S> isolateBloc,
-    IsolateBlocBuilderCondition<S> buildWhen,
-  })  : assert(builder != null),
-        super(key: key, isolateBloc: isolateBloc, buildWhen: buildWhen);
+    Key? key,
+    required this.builder,
+    IsolateBlocWrapper? bloc,
+    BlocBuilderCondition<S>? buildWhen,
+  }) : super(key: key, bloc: bloc, buildWhen: buildWhen);
 
   /// The [builder] function which will be invoked on each widget build.
   /// The [builder] takes the `BuildContext` and current `state` and
   /// must return a widget.
   /// This is analogous to the [builder] function in [StreamBuilder].
-  final IsolateBlocWidgetBuilder<S> builder;
+  final BlocWidgetBuilder<S> builder;
 
   @override
   Widget build(BuildContext context, S state) => builder(context, state);
@@ -95,94 +95,74 @@ class IsolateBlocBuilder<C extends IsolateBloc<Object, S>, S>
 
 /// {@template bloc_builder_base}
 /// Base class for widgets that build themselves based on interaction with
-/// a specified [isolateBloc].
+/// a specified [bloc].
 ///
 /// A [IsolateBlocBuilderBase] is stateful and maintains the state of the interaction
 /// so far. The type of the state and how it is updated with each interaction
 /// is defined by sub-classes.
 /// {@endtemplate}
-abstract class IsolateBlocBuilderBase<C extends IsolateBloc<Object, S>, S>
+abstract class IsolateBlocBuilderBase<B extends IsolateBlocBase<Object?, S>, S>
     extends StatefulWidget {
   /// {@macro bloc_builder_base}
-  const IsolateBlocBuilderBase({Key key, this.isolateBloc, this.buildWhen})
+  const IsolateBlocBuilderBase({Key? key, this.bloc, this.buildWhen})
       : super(key: key);
 
-  /// The [isolateBloc] that the [IsolateBlocBuilderBase] will interact with.
+  /// The [bloc] that the [IsolateBlocBuilderBase] will interact with.
   /// If omitted, [IsolateBlocBuilderBase] will automatically perform a lookup using
   /// [IsolateBlocProvider] and the current `BuildContext`.
-  final IsolateBlocWrapper<S> isolateBloc;
+  final IsolateBlocWrapper? bloc;
 
   /// {@macro bloc_builder_build_when}
-  final IsolateBlocBuilderCondition<S> buildWhen;
+  final BlocBuilderCondition<S>? buildWhen;
 
   /// Returns a widget based on the `BuildContext` and current [state].
   Widget build(BuildContext context, S state);
 
   @override
-  State<IsolateBlocBuilderBase<C, S>> createState() =>
-      _IsolateBlocBuilderBaseState<C, S>();
+  State<IsolateBlocBuilderBase<B, S>> createState() =>
+      _IsolateBlocBuilderBaseState<B, S>();
 }
 
-class _IsolateBlocBuilderBaseState<C extends IsolateBloc<Object, S>, S>
-    extends State<IsolateBlocBuilderBase<C, S>> {
-  StreamSubscription<S> _subscription;
-  S _previousState;
-  S _state;
-  IsolateBlocWrapper<S> _blocWrapper;
+class _IsolateBlocBuilderBaseState<B extends IsolateBlocBase<Object?, S>, S>
+    extends State<IsolateBlocBuilderBase<B, S>> {
+  late IsolateBlocWrapper _bloc;
+  late S _state;
 
   @override
   void initState() {
     super.initState();
-    _blocWrapper = widget.isolateBloc ?? context.isolateBloc<C, S>();
-    _previousState = _blocWrapper?.state;
-    _state = _blocWrapper?.state;
-    _subscribe();
+    _bloc = widget.bloc ?? context.isolateBloc<B, S>();
+    _state = _bloc.state!;
   }
 
   @override
-  void didUpdateWidget(IsolateBlocBuilderBase<C, S> oldWidget) {
+  void didUpdateWidget(IsolateBlocBuilderBase<B, S> oldWidget) {
     super.didUpdateWidget(oldWidget);
-    // ignore: close_sinks
-    final oldIsolateBloc = oldWidget.isolateBloc ?? context.isolateBloc<C, S>();
-    // ignore: close_sinks
-    final currentBloc = widget.isolateBloc ?? oldIsolateBloc;
-    if (oldIsolateBloc != currentBloc) {
-      if (_subscription != null) {
-        _unsubscribe();
-        _blocWrapper = widget.isolateBloc ?? context.isolateBloc<C, S>();
-        _previousState = _blocWrapper?.state;
-        _state = _blocWrapper?.state;
-      }
-      _subscribe();
+    final oldBloc = oldWidget.bloc ?? context.isolateBloc<B, S>();
+    final currentBloc = widget.bloc ?? oldBloc;
+    if (oldBloc != currentBloc) {
+      _bloc = currentBloc;
+      _state = _bloc.state!;
     }
   }
 
   @override
-  Widget build(BuildContext context) => widget.build(context, _state);
+  void didChangeDependencies() {
+    super.didChangeDependencies();
+    final bloc = widget.bloc ?? context.isolateBloc<B, S>();
+    if (_bloc != bloc) {
+      _bloc = bloc;
+      _state = _bloc.state!;
+    }
+  }
 
   @override
-  void dispose() {
-    _unsubscribe();
-    super.dispose();
-  }
-
-  void _subscribe() {
-    if (_blocWrapper != null) {
-      _subscription = _blocWrapper.listen((state) {
-        if (widget.buildWhen?.call(_previousState, state) ?? true) {
-          setState(() {
-            _state = state;
-          });
-        }
-        _previousState = state;
-      });
-    }
-  }
-
-  void _unsubscribe() {
-    if (_subscription != null) {
-      _subscription.cancel();
-      _subscription = null;
-    }
+  Widget build(BuildContext context) {
+    return IsolateBlocListener<B, S>(
+      bloc: _bloc,
+      listenWhen: widget.buildWhen,
+      listener: (context, state) => setState(() => _state = state),
+      child: widget.build(context, _state),
+    );
   }
 }
