@@ -1,7 +1,7 @@
 import 'package:http/http.dart';
 import 'package:isolate_bloc/isolate_bloc.dart';
-import 'package:weather_app/blocs/settings_bloc.dart';
-import 'package:weather_app/blocs/theme_bloc.dart';
+import 'package:weather_app/blocs/settings_cubit.dart';
+import 'package:weather_app/blocs/theme_cubit.dart';
 import 'package:weather_app/blocs/weather_bloc.dart';
 import 'package:weather_app/repositories/repositories.dart';
 
@@ -15,10 +15,13 @@ void isolateInitialization() {
     ),
   );
 
-  register(create: () => WeatherBloc(weatherRepository: weatherRepository));
-  register(create: () => SettingsBloc());
-  register(
-    create: () => ThemeBloc(
+  register<WeatherBloc, WeatherState>(
+    create: () => WeatherBloc(weatherRepository: weatherRepository),
+    initialState: WeatherInitial(),
+  );
+  register<SettingsCubit, SettingsState>(create: () => SettingsCubit());
+  register<ThemeCubit, ThemeState>(
+    create: () => ThemeCubit(
       weatherBloc: getBloc<WeatherBloc, WeatherState>(),
     ),
   );
