@@ -26,19 +26,19 @@ class Weather extends Equatable {
   final String location;
 
   const Weather({
-    this.condition,
-    this.formattedCondition,
-    this.minTemp,
-    this.temp,
-    this.maxTemp,
-    this.locationId,
-    this.created,
-    this.lastUpdated,
-    this.location,
+    required this.condition,
+    required this.formattedCondition,
+    required this.minTemp,
+    required this.temp,
+    required this.maxTemp,
+    required this.locationId,
+    required this.created,
+    required this.lastUpdated,
+    required this.location,
   });
 
   @override
-  List<Object> get props => [
+  List<Object?> get props => [
         condition,
         formattedCondition,
         minTemp,
