@@ -5,7 +5,7 @@ Future<void> main(List<String> arguments) async {
   await initialize(isolatedFunc);
   runApp(
     MaterialApp(
-      home: IsolateBlocProvider<CounterBloc, int>(
+      home: IsolateBlocProvider<CounterCubit, int>(
         child: CounterScreen(),
       ),
     ),
@@ -20,9 +20,9 @@ class CounterScreen extends StatelessWidget {
         title: const Text('Counter'),
       ),
       body: Center(
-        child: IsolateBlocListener<CounterBloc, int>(
+        child: IsolateBlocListener<CounterCubit, int>(
           listener: (context, state) => print('New bloc state: $state'),
-          child: IsolateBlocBuilder<CounterBloc, int>(
+          child: IsolateBlocBuilder<CounterCubit, int>(
             builder: (context, state) {
               return Text('You tapped $state times');
             },
@@ -35,7 +35,7 @@ class CounterScreen extends StatelessWidget {
           FloatingActionButton(
             heroTag: 'Increment',
             onPressed: () => context
-                .isolateBloc<CounterBloc, int>()
+                .isolateBloc<CounterCubit, int>()
                 .add(CountEvent.increment),
             child: const Icon(Icons.add),
           ),
@@ -43,7 +43,7 @@ class CounterScreen extends StatelessWidget {
           FloatingActionButton(
             heroTag: 'Decrement',
             onPressed: () => context
-                .isolateBloc<CounterBloc, int>()
+                .isolateBloc<CounterCubit, int>()
                 .add(CountEvent.decrement),
             child: const Icon(Icons.remove),
           ),
@@ -55,11 +55,11 @@ class CounterScreen extends StatelessWidget {
 
 Future<void> isolatedFunc() async {
   IsolateBloc.observer = SimpleBlocObserver();
-  register(create: () => CounterBloc());
+  register<CounterCubit, int>(create: () => CounterCubit());
 }
 
-class CounterBloc extends IsolateBloc<CountEvent, int> {
-  CounterBloc() : super(0);
+class CounterCubit extends IsolateCubit<CountEvent, int> {
+  CounterCubit() : super(0);
 
   @override
   void onEventReceived(CountEvent event) {
@@ -74,7 +74,13 @@ enum CountEvent {
 
 class SimpleBlocObserver extends IsolateBlocObserver {
   @override
-  void onEvent(IsolateBloc bloc, Object event) {
+  void onClose(IsolateBlocBase bloc) {
+    print('New instance of ${bloc.runtimeType}');
+    super.onClose(bloc);
+  }
+
+  @override
+  void onEvent(IsolateBlocBase bloc, Object? event) {
     print('New $event for $bloc');
     super.onEvent(bloc, event);
   }
@@ -86,7 +92,7 @@ class SimpleBlocObserver extends IsolateBlocObserver {
   }
 
   @override
-  void onError(IsolateBloc bloc, Object error, StackTrace stackTrace) {
+  void onError(IsolateBlocBase bloc, Object error, StackTrace stackTrace) {
     print('$error in $bloc');
     super.onError(bloc, error, stackTrace);
   }
