@@ -2,16 +2,21 @@ library isolate_bloc;
 
 export 'src/common/api_wrappers.dart';
 export 'src/common/bloc/isolate_bloc.dart';
+export 'src/common/bloc/isolate_bloc_base.dart';
 export 'src/common/bloc/isolate_bloc_observer.dart';
 export 'src/common/bloc/isolate_bloc_wrapper.dart';
+export 'src/common/bloc/isolate_cubit.dart';
 export 'src/common/bloc/transition.dart';
-export 'src/common/isolate/bloc_manager.dart';
-export 'src/common/isolate/isolate_manager/abstract_isolate_manager.dart';
-export 'src/common/isolate/isolate_manager/abstract_isolate_wrapper.dart';
-export 'src/common/isolate/isolate_manager/isolate_messenger.dart';
-export 'src/common/isolate/platform_channel/platform_channel_setup.dart';
+export 'src/common/isolate/manager/isolate_manager.dart';
+export 'src/common/isolate/manager/ui_isolate_manager.dart';
+export 'src/common/isolate/initializer/isolate_initializer.dart';
+export 'src/common/isolate/isolate_factory/i_isolate_factory.dart';
+export 'src/common/isolate/isolate_factory/i_isolate_wrapper.dart';
+export 'src/common/isolate/isolate_factory/isolate_messenger/isolate_messenger.dart';
+export 'src/common/isolate/method_channel/method_channel_setup.dart';
+export 'src/common/isolated_api_wrappers.dart';
 export 'src/flutter/isolate_bloc_builder.dart';
 export 'src/flutter/isolate_bloc_consumer.dart';
 export 'src/flutter/isolate_bloc_listener.dart';
-export 'src/flutter/isolate_bloc_multi_provider.dart';
+export 'src/flutter/multi_isolate_bloc_provider.dart';
 export 'src/flutter/isolate_bloc_provider.dart';
