@@ -1,195 +1,185 @@
 import 'dart:async';
 
 import 'package:flutter/foundation.dart';
-import 'package:isolate_bloc/isolate_bloc.dart';
-import 'package:uuid/uuid.dart';
+import 'package:isolate_bloc/src/common/bloc/isolate_bloc_base.dart';
+import 'package:isolate_bloc/src/common/bloc/isolate_bloc_observer.dart';
+import 'package:isolate_bloc/src/common/bloc/transition.dart';
 
-import 'isolate_bloc_observer.dart';
-import 'transition.dart';
+/// Signature for a mapper function which takes an [Event] as input
+/// and outputs a [Stream] of [Transition] objects.
+typedef TransitionFunction<Event, State> = Stream<Transition<Event, State>>
+    Function(Event);
 
-class BlocUnhandledErrorException implements Exception {
-  BlocUnhandledErrorException(this.bloc, this.error, [this.stackTrace]);
-
-  /// The [bloc] in which the unhandled error occurred.
-  final IsolateBloc bloc;
-
-  /// The unhandled [error] object.
-  final Object error;
-
-  /// An optional [stackTrace] which accompanied the error.
-  final StackTrace stackTrace;
-
-  @override
-  String toString() {
-    return 'Unhandled error $error occurred in bloc $bloc.\n'
-        '${stackTrace ?? ''}';
-  }
-}
-
-/// This bloc works in isolate.
-/// [IsolateBlocWrapper]<Event>, State> will receive this bloc's state.
-/// This bloc must be created from UI with `createBloc<BlocT, BlocTState>`
-/// function or `IsolateBlocProvider<BlocT, BlocTState>()`.
-/// You can use it from another [IsolateBloc] with `getBloc<BlocT>()`
-/// or `getBlocWrapper<BlocT, BlocTState>()`.
-abstract class IsolateBloc<Event, State> extends Stream<State>
-    implements Sink<State> {
-  /// Basic constructor. Gain initial state and generate bloc's id;
-  IsolateBloc(this._state) : id = Uuid().v4() {
-    _bindStateReceiver();
+/// {@macro isolate_cubit_description}
+abstract class IsolateBloc<Event, State> extends IsolateBlocBase<Event, State> {
+  /// {@macro isolate_cubit_description}
+  IsolateBloc(State state) : super(state) {
+    _bindEventsToStates();
   }
 
-  State _state;
-  Event _event;
-
-  /// This is bloc's id. Every [IsolateBloc] have it's own unique id used to
-  /// communicate with it's own [IsolateBlocWrapper].
-  final String id;
-  final _stateController = StreamController<State>.broadcast();
-  StreamSubscription<State> _stateSubscription;
-
-  /// The current [IsolateBlocObserver].
+  /// The current [IsolateBlocObserver] instance.
   static IsolateBlocObserver observer = IsolateBlocObserver();
 
-  /// Returns the current [state] of the [IsolateBloc].
-  State get state => _state;
+  StreamSubscription<Transition<Event, State>>? _transitionSubscription;
 
-  /// Notifies the [IsolateBloc] of a new event which triggers onEventReceived.
-  @mustCallSuper
+  final _eventController = StreamController<Event>.broadcast();
+
+  /// Transforms the [events] stream along with a [transitionFn] function into
+  /// a `Stream<Transition>`.
+  /// Events that should be processed by [mapEventToState] need to be passed to
+  /// [transitionFn].
+  /// By default `asyncExpand` is used to ensure all [events] are processed in
+  /// the order in which they are received.
+  /// You can override [transformEvents] for advanced usage in order to
+  /// manipulate the frequency and specificity with which [mapEventToState] is
+  /// called as well as which [events] are processed.
+  ///
+  /// For example, if you only want [mapEventToState] to be called on the most
+  /// recent [Event] you can use `switchMap` instead of `asyncExpand`.
+  ///
+  /// ```dart
+  /// @override
+  /// Stream<Transition<Event, State>> transformEvents(events, transitionFn) {
+  ///   return events.switchMap(transitionFn);
+  /// }
+  /// ```
+  ///
+  /// Alternatively, if you only want [mapEventToState] to be called for
+  /// distinct [events]:
+  ///
+  /// ```dart
+  /// @override
+  /// Stream<Transition<Event, State>> transformEvents(events, transitionFn) {
+  ///   return super.transformEvents(
+  ///     events.distinct(),
+  ///     transitionFn,
+  ///   );
+  /// }
+  /// ```
+  Stream<Transition<Event, State>> transformEvents(
+    Stream<Event> events,
+    TransitionFunction<Event, State> transitionFn,
+  ) {
+    return events.asyncExpand(transitionFn);
+  }
+
+  /// {@template emit}
+  /// **[emit] should never be used outside of tests.**
+  ///
+  /// Updates the state of the bloc to the provided [state].
+  /// A bloc's state should only be updated by `yielding` a new `state`
+  /// from `mapEventToState` in response to an event.
+  /// {@endtemplate}
+  @protected
+  @visibleForTesting
   @override
-  void add(Object event) {
-    try {
-      _event = event as Event;
-      onEvent(_event);
-      onEventReceived(_event);
-    } catch (e, stackTrace) {
-      onError(e, stackTrace);
+  void emit(State state) => super.emit(state);
+
+  /// Called whenever new [event] received
+  ///
+  /// Can't be overridden because must add [event] to the [_eventController]
+  @protected
+  @override
+  void onEventReceived(Event event) {
+    if (!_eventController.isClosed) {
+      _eventController.add(event);
     }
   }
 
-  /// Use this function to send new state to UI.
-  @mustCallSuper
-  void emit(State state) {
-    _stateController.add(state);
-  }
+  /// Must be implemented when a class extends [IsolateBloc].
+  /// [mapEventToState] is called whenever an [event] is [add]ed
+  /// and is responsible for converting that [event] into a new [state].
+  /// [mapEventToState] can `yield` zero, one, or multiple states for an event.
+  Stream<State> mapEventToState(Event event);
 
   /// Called whenever a [transition] occurs with the given [transition].
-  /// A [transition] occurs when a new `event` is [add]ed and `state`
-  /// from `emit` received.
-  /// [onTransition] is called before a [bloc]'s [state] has been updated.
-  /// A great spot to add logging/analytics at the individual [bloc] level.
+  /// A [transition] occurs when a new `event` is [add]ed and [mapEventToState]
+  /// executed.
+  /// [onTransition] is called before a [IsolateBloc]'s [state] has been updated.
+  /// A great spot to add logging/analytics at the individual [IsolateBloc] level.
   ///
-  /// **Note: `super.onTransition` should always be called last.**
+  /// **Note: `super.onTransition` should always be called first.**
   /// ```dart
   /// @override
   /// void onTransition(Transition<Event, State> transition) {
-  ///   // Custom onTransition logic goes here
-  ///
   ///   // Always call super.onTransition with the current transition
   ///   super.onTransition(transition);
+  ///
+  ///   // Custom onTransition logic goes here
   /// }
   /// ```
+  ///
+  /// See also:
+  ///
+  /// * [IsolateBlocObserver.onTransition] for observing transitions globally.
+  ///
+  @protected
   @mustCallSuper
-  void onTransition(Transition<Object, State> transition) {
+  void onTransition(Transition<Event, State> transition) {
     // ignore: invalid_use_of_protected_member
     IsolateBloc.observer.onTransition(this, transition);
   }
 
-  /// Called whenever an [event] is [add]ed to the [bloc].
-  /// A great spot to add logging/analytics at the individual [bloc] level.
+  /// Transforms the `Stream<Transition>` into a new `Stream<Transition>`.
+  /// By default [transformTransitions] returns
+  /// the incoming `Stream<Transition>`.
+  /// You can override [transformTransitions] for advanced usage in order to
+  /// manipulate the frequency and specificity at which `transitions`
+  /// (state changes) occur.
   ///
-  /// **Note: `super.onEvent` should always be called last.**
+  /// For example, if you want to debounce outgoing state changes:
+  ///
   /// ```dart
   /// @override
-  /// void onEvent(Event event) {
-  ///   // Custom onEvent logic goes here
-  ///
-  ///   // Always call super.onEvent with the current event
-  ///   super.onEvent(event);
+  /// Stream<Transition<Event, State>> transformTransitions(
+  ///   Stream<Transition<Event, State>> transitions,
+  /// ) {
+  ///   return transitions.debounceTime(Duration(seconds: 1));
   /// }
   /// ```
-  @mustCallSuper
-  void onEvent(Object event) {
-    // ignore: invalid_use_of_protected_member
-    IsolateBloc.observer.onEvent(this, event);
+  Stream<Transition<Event, State>> transformTransitions(
+    Stream<Transition<Event, State>> transitions,
+  ) {
+    return transitions;
   }
 
-  /// Called whenever an [error] is thrown within onTransition in [IsolatedBloc].
-  /// By default all [error]s will be ignored and [bloc] functionality will be
-  /// unaffected.
-  /// The [stackTrace] argument may be `null` if the [state] stream received
-  /// an error without a [stackTrace].
-  /// A great spot to handle errors at the individual [IsolateBlocWrapper] level.
-  ///
-  /// **Note: `super.onError` should always be called last.**
-  /// ```dart
-  /// @override
-  /// void onError(Object error, StackTrace stackTrace) {
-  ///   // Custom onError logic goes here
-  ///
-  ///   // Always call super.onError with the current error and stackTrace
-  ///   super.onError(error, stackTrace);
-  /// }
-  /// ```
-  @mustCallSuper
-  void onError(Object error, StackTrace stackTrace) {
-    // ignore: invalid_use_of_protected_member
-    IsolateBloc.observer.onError(this, error, stackTrace);
-    if (kDebugMode) {
-      throw BlocUnhandledErrorException(this, error, stackTrace);
-    }
-  }
-
-  /// Adds a subscription to the Stream<State>. Returns a StreamSubscription
-  /// which handles events from the Stream<State> using the provided onData,
-  /// onError and onDone handlers.
-  @override
-  StreamSubscription<State> listen(
-    void Function(State event) onData, {
-    Function onError,
-    void Function() onDone,
-    bool cancelOnError,
-  }) {
-    return _prepareStateStream.listen(
-      onData,
-      onError: onError,
-      onDone: onDone,
-      cancelOnError: cancelOnError,
-    );
-  }
-
-  Stream<State> get _prepareStateStream async* {
-    yield state;
-    yield* _stateController.stream;
-  }
-
-  /// This function receive events from UI.
-  void onEventReceived(Event event);
-
-  void _bindStateReceiver() {
-    _stateSubscription = _stateController.stream.listen((nextState) {
-      final transition = Transition(
-        currentState: state,
-        event: _event,
-        nextState: nextState,
-      );
-
-      try {
-        onTransition(transition);
-        _state = transition.nextState;
-      } on dynamic catch (error, stackTrace) {
-        onError(error, stackTrace);
-      }
-    });
-  }
-
-  /// Free all resources. This method should be called when a IsolateBloc is no
-  /// longer needed. Once close is called, events that are added will not be
+  /// Closes the `event` and `state` `Streams`.
+  /// This method should be called when a [IsolateBloc] is no longer needed.
+  /// Once [close] is called, `events` that are [add]ed will not be
   /// processed.
+  /// In addition, if [close] is called while `events` are still being
+  /// processed, the [IsolateBloc] will finish processing the pending `events`.
   @override
   @mustCallSuper
   Future<void> close() async {
-    await _stateSubscription.cancel();
-    await _stateController.close();
+    await _transitionSubscription?.cancel();
+
+    return super.close();
+  }
+
+  void _bindEventsToStates() {
+    _transitionSubscription = transformTransitions(
+      transformEvents(
+        _eventController.stream,
+        (event) => mapEventToState(event).map(
+          (nextState) => Transition(
+            currentState: state,
+            event: event,
+            nextState: nextState,
+          ),
+        ),
+      ),
+    ).listen(
+      (transition) {
+        if (transition.nextState == state && emitted) return;
+        try {
+          onTransition(transition);
+          emit(transition.nextState);
+        } catch (error, stackTrace) {
+          onError(error, stackTrace);
+        }
+      },
+      onError: onError,
+    );
   }
 }
