@@ -1,11 +1,11 @@
 import 'package:isolate_bloc/isolate_bloc.dart';
 
-import 'blocs/simple_bloc.dart';
+import 'blocs/simple_cubit.dart';
 
 class Initializers {
   static void empty() {}
 
   static void simpleBloc() {
-    register(create: () => SimpleBloc());
+    register<SimpleCubit, Object>(create: () => SimpleCubit());
   }
 }
